Rename CClaimTrieCache to CClaimTrieUpdateBuffer and separate in own file
