+++ src/Makefile.test.include
@@ -61,7 +61,7 @@ BITCOIN_TESTS =\
   test/miner_tests.cpp \
   test/multisig_tests.cpp \
   test/claimtrie_tests.cpp \
-  test/claimtriecache_tests.cpp \
+  test/claimtrieupdatebuffer_tests.cpp \
   test/claimtriebranching_tests.cpp \
   test/claimtriedb_tests.cpp \
   test/nameclaim_tests.cpp \
